refactor: seasonal 데이터 import 경로 업데이트
