+++ src/components/Question.jsx
@@ -3,7 +3,7 @@ import { useParams, useNavigate } from "react-router-dom";
 import styled from "styled-components";
 import axios from "axios";
 import { agree, scheduleData, required } from "../data/questions/questionData";
-import { year } from "../data/seasonal/seasonalData";
+import { year } from "../data/seasonal/aboutus";
 
 const Question = () => {
   const { part } = useParams();
